backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
langgraph>=0.0.40
langchain-core>=0.1.0
typing_extensions>=4.0.0
gitpython>=3.1.40
pydantic>=2.5.0
httpx>=0.26.0
python-dotenv>=1.0.0
